+++ packages/flame/lib/src/components/parallax_component.dart
@@ -12,16 +12,17 @@ import 'position_component.dart';
 
 extension ParallaxComponentExtension on Game {
   Future<ParallaxComponent> loadParallaxComponent(
-    List<String> paths, {
+    List<ParallaxData> dataList, {
     Vector2? size,
     Vector2? baseVelocity,
     Vector2? velocityMultiplierDelta,
     ImageRepeat repeat = ImageRepeat.repeatX,
     Alignment alignment = Alignment.bottomLeft,
     LayerFill fill = LayerFill.height,
+    int? priority,
   }) async {
     final component = await ParallaxComponent.load(
-      paths,
+      dataList,
       size: size,
       baseVelocity: baseVelocity,
       velocityMultiplierDelta: velocityMultiplierDelta,
@@ -29,6 +30,7 @@ extension ParallaxComponentExtension on Game {
       alignment: alignment,
       fill: fill,
       images: images,
+      priority: priority,
     );
 
     return component;
@@ -51,7 +53,8 @@ class ParallaxComponent extends PositionComponent {
   ParallaxComponent({
     Vector2? position,
     Vector2? size,
-  }) : super(position: position, size: size) {
+    int? priority,
+  }) : super(position: position, size: size, priority: priority) {
     if (size != null) {
       isFullscreen = false;
     }
@@ -61,10 +64,12 @@ class ParallaxComponent extends PositionComponent {
   factory ParallaxComponent.fromParallax(
     Parallax parallax, {
     Vector2? position,
+    int? priority,
   }) {
     return ParallaxComponent(
       position: position,
       size: parallax.isSized ? parallax.size : null,
+      priority: priority,
     )..parallax = parallax;
   }
 
@@ -96,8 +101,9 @@ class ParallaxComponent extends PositionComponent {
   /// and filled), otherwise load the [ParallaxLayer]s individually and use the
   /// normal constructor.
   ///
-  /// [load] takes a list of paths to all the images and a size that you want to use in the
+  /// [load] takes a list of [ParallaxData] of all the images and a size that you want to use in the
   /// parallax.
+  ///
   /// Optionally arguments for the [baseVelocity] and [velocityMultiplierDelta] can be passed
   /// in, [baseVelocity] defines what the base velocity of the layers should be
   /// and [velocityMultiplierDelta] defines how the velocity should change the
@@ -107,9 +113,10 @@ class ParallaxComponent extends PositionComponent {
   /// which edge it should align with ([alignment]), which axis it should fill
   /// the image on ([fill]) and [images] which is the image cache that should be
   /// used can also be passed in.
+  ///
   /// If no image cache is set, the global flame cache is used.
   static Future<ParallaxComponent> load(
-    List<String> paths, {
+    List<ParallaxData> dataList, {
     Vector2? size,
     Vector2? baseVelocity,
     Vector2? velocityMultiplierDelta,
@@ -117,10 +124,11 @@ class ParallaxComponent extends PositionComponent {
     Alignment alignment = Alignment.bottomLeft,
     LayerFill fill = LayerFill.height,
     Images? images,
+    int? priority,
   }) async {
     final component = ParallaxComponent.fromParallax(
       await Parallax.load(
-        paths,
+        dataList,
         size: size,
         baseVelocity: baseVelocity,
         velocityMultiplierDelta: velocityMultiplierDelta,
@@ -129,6 +137,7 @@ class ParallaxComponent extends PositionComponent {
         fill: fill,
         images: images,
       ),
+      priority: priority,
     );
 
     if (size != null) {