+++ packages/flame/lib/src/particles/circle_particle.dart
@@ -2,9 +2,9 @@ import 'dart:ui';
 
 import 'particle.dart';
 
-/// Plain circle with no other behaviors
-/// Consider composing with other [Particle]
-/// to achieve needed effects
+/// Plain circle with no other behaviors.
+///
+/// Consider composing this with other [Particle]s to achieve needed effects.
 class CircleParticle extends Particle {
   final Paint paint;
   final double radius;
@@ -13,9 +13,7 @@ class CircleParticle extends Particle {
     required this.paint,
     this.radius = 10.0,
     double? lifespan,
-  }) : super(
-          lifespan: lifespan,
-        );
+  }) : super(lifespan: lifespan);
 
   @override
   void render(Canvas c) {